pcg/requirements.txt
Pillow >= 4.0.0